feat: remove succeedWith and failWith.
