+++ packages/pure-parse/src/parsers/index.ts
@@ -9,4 +9,4 @@ export * from './optional'
 export * from './instanceOf'
 export * from './never'
 export * from './unknown'
-export * from './defaults'
+export * from './withDefault'
