+++ packages/pure-parse/src/parsers/withDefault.test.ts
@@ -1,0 +1,49 @@
+import { describe, expect, it, test } from 'vitest'
+import { withDefault } from './withDefault'
+import { parseString } from './primitives'
+import { object } from './object'
+
+describe('withDefault', () => {
+  test('as a root parser', () => {
+    const parseStr = withDefault(parseString, 'fallback')
+    expect(parseStr('hello')).toEqual(
+      expect.objectContaining({
+        value: 'hello',
+      }),
+    )
+    expect(parseStr(123)).toEqual(
+      expect.objectContaining({
+        value: 'fallback',
+      }),
+    )
+  })
+  test('as a nested parser', () => {
+    const parseObj = object({
+      name: withDefault(parseString, 'fallback'),
+    })
+    expect(parseObj({ name: 'hello' })).toEqual(
+      expect.objectContaining({
+        value: { name: 'hello' },
+      }),
+    )
+  })
+  describe('fallbackValue on fallbackValue', () => {
+    it('uses the first fallbackValue', () => {
+      const parseStr = withDefault(
+        withDefault(parseString, 'fallback'),
+        'fallback2',
+      )
+      expect(parseStr('hello')).toEqual(
+        expect.objectContaining({
+          value: 'hello',
+        }),
+      )
+      expect(parseStr(123)).toEqual(
+        expect.objectContaining({
+          value: 'fallback',
+        }),
+      )
+    })
+  })
+  test.todo('on optional properties')
+})
